requirements.txt
alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
click==8.1.7
flasgger==0.9.7.1
flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-Script==2.0.6
flask-sqlalchemy==3.1.1
flask-swagger-ui==4.11.1
flask-wtf==1.2.1
greenlet==3.0.3
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.5
mistune==3.0.2
packaging==23.2
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
six==1.16.0
SQLAlchemy==2.0.25
typing-extensions==4.9.0
werkzeug==3.0.1
wtforms==3.1.2
zipp==3.17.0
